Complete ch04

+++ ch03/Account/Account.h
@@ -6,7 +6,7 @@
 class Account {
 public:
    // Account constructor with two parameters  
-   Account(std::string accountName, int initialBalance) 
+   Account(std::string accountName, double initialBalance) 
       : name{accountName} { // assign accountName to data member name
 
       // validate that the initialBalance is greater than 0; if not,
@@ -17,7 +17,7 @@ public:
    }
 
    // function that deposits (adds) only a valid amount to the balance
-   void deposit(int depositAmount) {
+   void deposit(double depositAmount) {
       if (depositAmount > 0) { // if the depositAmount is valid
          balance = balance + depositAmount; // add it to the balance 
       }
@@ -25,14 +25,14 @@ public:
 
 
     // function that withdraw (remove) only a valid amount to the balance
-   void withdraw(int withdrawAmount) {
+   void withdraw(double withdrawAmount) {
       if (withdrawAmount <= balance) { // if the withdrawAmount is valid
          balance = balance - withdrawAmount; // remove it to the balance 
       }
    }
 
    // function returns the account balance
-   int getBalance() const {
+   double getBalance() const {
       return balance;
    }
 
@@ -47,7 +47,7 @@ public:
    }
 private:
    std::string name; // account name data member 
-   int balance{0}; // data member with default initial value
+   double balance{0}; // data member with default initial value
 }; // end class Account
